watchdog support for spark-3.2
