+++ dev/kyuubi-extension-spark-3-2/src/test/scala/org/apache/spark/sql/WatchDogSuite.scala
@@ -1,0 +1,3 @@
+package org.apache.spark.sql
+
+class WatchDogSuite extends WatchDogSuiteBase {}
